+++ wpull/options.py
@@ -819,6 +819,12 @@ class AppArgumentParser(argparse.ArgumentParser):
         if hasattr(ssl, 'PROTOCOL_SSLv2'):
             self._ssl_version_map['SSLv2'] = ssl.PROTOCOL_SSLv2,
 
+        if hasattr(ssl, 'PROTOCOL_TLSv1_1'):
+            self._ssl_version_map['TLSv1.1'] = ssl.PROTOCOL_TLSv1_1,
+
+        if hasattr(ssl, 'PROTOCOL_TLSv1_2'):
+            self._ssl_version_map['TLSv1.2'] = ssl.PROTOCOL_TLSv1_2,
+
         group = self.add_argument_group('SSL')
         group.add_argument(
             '--secure-protocol',
@@ -838,6 +844,13 @@ class AppArgumentParser(argparse.ArgumentParser):
             action='store_false',
             default=True,
             help=_('don’t validate SSL server certificates'),
+        )
+        group.add_argument(
+            '--no-strong-crypto',
+            dest='strong_crypto',
+            action='store_false',
+            default=True,
+            help=_('don’t use secure protocols/ciphers')
         )
         group.add_argument(
             '--certificate',