+++ doc/changelog.rst
@@ -2,6 +2,7 @@
 What's New
 ==========
 
+* Security: SSLv2/SSLv3 is disabled for ``--secure-protocol=auto``. Added ``--no-strong-crypto`` that re-enables them again if needed.
 * Fixed NameError with PhantomJS proxy on Python 3.2.
 * Fixed PhantomJS stop waiting for page load too early.
 * Fixed ``--page-requisites`` exceeding ``--level``.